quick poc for Thalhammer#174
